Only posts with titles will show on Archives page
Removing references to blog category as the lack of a Title flags this for us.

Short posts are considered ephemeral and will not show in the archive even if they have photos in them.

Photos will only show in the archive if they are part of a titled post.

Updated the About page accordingly